// reality/cloud/xrhome/src/client/common/types/db.ts
type AppHostingType = "UNSET" | "CLOUD_EDITOR" | "CLOUD_STUDIO" | "SELF" | "INTERNAL" | "AD"